requirements.txt
discord.py>=1.5.1
python-dotenv==0.19.1
Pillow==8.3.2